refactor(bayn): modularize paper candidate discovery
